framework/settings: added a debug mode (--debug or -d option).
The debug mode is disabled by default and may be enabled through the -d
or --debug command line arguments or by defining the 'debug' setting in
the configuration file. Currently, when the debug mode is enabled, the
framework prints information needed by the GDB debugger to locate the
program analyser library (.so file) loaded by the framework.