framework/settings: the ANaConDA framework's library is now loaded twice.
The ANaConDA framework's library is now loaded by the framework itself
before loading the analyser. This ensures that the exported symbols are
accessible to the analyser. For some unknown reason, when the framework
is loaded only by the PIN framework, its symbols are not accessible to
the program analyser and everything crashes on unresolved symbol error.