framework/mapper: fixed a data race(s) on the index map (and index counter).
Because mapping system-specific locks and conditions to corresponding
lock and condition objects happen in parallel, there is a possible data
race on the index map and index counter attributes of the mapper object.
This data race occurs very rarely (once in more than 10 000 runs), but
still might occur. Fixed using an RW mutex.