framework/access: callback functions are now instances of two template functions.
The two template functions are able to generate all the needed callback
functions. This approach allows to share the code between several types
of callback functions simplifying the maintenance while still retaining
the performance of the previous implementation. Based on the parameters
of the template function the compiler generates the minimal code needed
to process the memory access and provide information the user requested.