framework: REP instructions are now monitored correctly (temporary fix).
Before and after callback functions are now called in each iteration of
a REP sequence. Using the InsertPredicatedCall function leads to a very
strange behaviour when sometimes no analysis functions are called after
a REP instruction is executed for the first time (i.e. after the first
iteration). For now, the framework performs all the needed checks, but
this is most likely a bug in the PIN framework which needs to be fixed.