framework/noise: all noise injection functions are now instanced from a template.
Added a template function which is able to generate all built-in noise
injection functions. The compiler generates the minimal code needed to
inject a noise of some type, so no performance degradation compared to
the previous implementation, but the maintenance is far simpler as the
code shared by all functions is at one location.