framework: backtraces of threads now created by monitoring base pointers.
Backtraces are now created by monitoring PUSH and POP instructions which
operate with base pointers and LEAVE instructions (which always operate
with base pointers). This requires the program to use the frame pointers
or it will be impossible to track the chain of base pointers to unwind
the stack and get the return addresses needed to create a backtrace.