libdie-wrapper/build: CMake now provides path and symbol information on Windows.
As automatic discovery of include paths and defined symbols seems not to
work in Eclipse on Windows when MSVC++ compiler is used, the information
are now provided by CMake. CMake exports this information to a text file
in a format which Eclipse scanners understand and Eclipse then loads the
content of the file to obtain this information.