framework/noise: overhauled the noise injection configuration system.
NoiseDesc transformed into NoiseSettings which holds the noise injection
settings for various types of locations. The settings are divided into 2
parts: a set of filters determining before which locations a noise might
be injected and a generator (a function generating noise) parameterized
by frequency and strength which determines if the noise should really be
injected and how strong it should be. This does not correspond with the
noise placement and noise seeding problems, because it is more effective
to do the frequency checks in the generator itself then as a part of the
filters where much more information must be passed. When no filters are
used, there is no need to call the filter function which speeds up the
execution a little and there is no reason to do not use the frequency
checks so there is no need to implement them as a separate filter.