framework/noise: inverse noise split into the generator and filter functions.
This allows one to setup the inverse noise in various ways. For example,
one may activate the inverse filter only for some locations blocking the
threads only before these locations and, e.g., injecting other types of
noise before the remaining locations etc. Also note that the position of
the filter in the filter list is important as the filters are called in
the same order as specified in the filter list. For example, placing the
filter before the shared variables filter will block the threads before
any access, however, placing it after the shared variables filter would
mean that the threads will be blocked only before accesses to shared
variables (because when a filter returns false, the remaining filters
are not evaluated).