framework/thread: added a new internal function addThreadInitFunction().
This function registers functions to be called when a thread start its
execution. Compared to the PIN_AddThreadStartFunction(), this function
might be called at any time, not only after PIN_Init() making it more
usable and also more efficient (we register only a single function in
PIN and in this function call all the registered functions at once, no
need to switch between PIN and analysis functions several times).