framework: moved the addThreadInitFunction() to a separate header.
The addThreadInitFunction() function was moved to a separate header to
prevent cyclic dependencies of headers. The remaining utility functions
still present in the callbacks/thread.h header will be moved to the new
utils/thread.h header in the future to prevent similar problems and to
separate the callback functions from the utility functions as they are
often used separately so there is no need to have them in one header.