framework/coverage: sync coverage monitor implementation moved to sync.hpp.
As the monitor is a template class, moving the whole implementation to a
header file makes more sense. There is no need to explicitly instantiate
the class in the source file, however, the compilation will be slower as
the monitor will not be compiled as a separate module now. In the future
precompiled headers might be utilized to deal with this problem.