framework/callbacks: callback functions now have HookType* as the last parameter.
It looks like it is safe to pass any pointer as a VOID pointer and then
get it in the callback function as another, e.g. HookType, pointer. So
replacing the VOID* parameter with a HookType* parameter should not
cause any problems and we do not need to static_cast the parameter.