framework/sync: fixed ThreadLocalData< ThreadData > to instantiate properly.
When instantiating ThreadLocalData class with the internal ThreadData
structure, the ThreadData structure must be in an anonymous namespace
in order for the name of the structure to be unique. If instantiating
the ThreadLocalData with two different structures of the same name
(e.g. using the internal ThreadData structures of various modules),
the instances which use different structures will use the same methods
for their management leading to severe error like segmentation faults.
This is because the structures might be internal and visible only to a
specific module, but when used by the ThreadLocalData class, they will
be used to identify the static methods generated by the class. If the
structures have the same name, the generated static methods will have
the same signature and the linker (or compiler) will leave only one of
these methods in the compiled binary and one of the instances of the
class will use wrong methods for managing the encapsulated data, often
causing errors like segmentation faults. Placing the internal structures
into an anonymous namespace will also make their name unique as the
compiler will mangle their name in order to protect them from accesses
from other modules.