framework: uses FunctionEntered/Exited callbacks to monitor predecessors.
No need to call the predecessors directly from the function execution
monitoring code as we have now the FunctionEntered/Exited callbacks,
which are triggered at the same time and provide the same information
the predecessor monitor needs. However, using the new callbacks leads
to a much simpler, more generic and more maintainable code.