framework/cbstack: after callbacks trigger correctly when performing a long jump.
Added support for triggering after callbacks when the functions do not
return because of the long jump. The framework now detects that a long
jump is performed and discovers which functions finish their execution
because of that. For each of the discovered functions, after callbacks
are properly triggered as if the functions had just returned (however,
there is no return value in this case, so the 0 address is returned as
the address of the return value).