framework/tm: added the Tx management operation result to its after callbacks.
After callbacks of the transaction start/commit/abort operations now
contain information about the result of the operation. The result is
provided as a reference to the return value to allow the analyser to
interpret it appropriately. If the operation did not return, because
of a long jump, for example, the reference will be NULL.