framework: noise points are not instrumented if noise frequency is zero.
If a noise which should be inserted before a noise point (function) has
a zero frequency, there is no need to insert the noise-generating code
into the program (as it will not ever be executed). However, inserting
the noise-generating code would introduce some noise into the program
when we do not want to do it (and set the zero frequency because of it).