tools/build: added support for building libdwarf library, few fixes.
Fixed update_env_var() to export all variables updated as some of them
are needed by the processes executed from the script (like CMake which
is used to perform some checks). Updated check_libdwarf() to check the
path to the libdwarf library instead of checking that libdwarf was not
found. Added missing CMake files needed to perform the libdwarf check.