tools/evaluate: added support for registering evaluators for each program.
An evaluator is a set of callback functions able to process the results
of a test. Each test consists of an arbitrary number of test runs where
the program under test is executed. There are three types of callback
functions: called before the evatuation, when processing each of the
test runs, and after the evaluation.