tools/update: access the server non-interactively, else use --interactive.
By default, the remote server is access using a standard bash shell,
i.e., non-login and non-interactive shell. This is because the files
to update definitions currently use only the variables from anaconda
environment, which are always available. No need to use interactive
connections to ensure that the standard environment variables are
properly set. Still, the old behaviour might be enforced using the
--interactive option.