tools/test: logs are now validated after the test is finished.
The log validation function checks the logs contaning information about
the test runs, processes this information, and creates a test.log file
which contains information about the outcome of each of the test runs.
The outcome might be succeeded (no errors), timeouted (test run stopped
because the time given to it runs out), or failed (some error occured).
If no entry is present for a given test run, the script was not able to
get the exit code of the tested program. This should happen only in one
case. That is, when the test run is interrupted because the time given
to the whole test runs out and the test script is stopped. In this case
the test run cannot be considered as timeouted nor failed, but also not
succeeded, so no entry in the test.log is present for it. This should
be the last test run performed and the results should no be taken into
account in this case.