tools/evaluate: added support for test run validation before their evaluation.
Evaluators may now optionally register a test run validation function
which may decide (based on the output of the test run) if the test run
is valid or not. Timeouted and failed test runs are not validated for
now as they are assumed to be invalid to begin with. The validation is
mainly for the cases when a test run ends without an error (with exit
status 0) despite the fact that error occured. These situations should
be very rare, but still might happend. In this case, the evaluator of
the program might tell the evaluation script, that such situation just
occured and force the script to skip the evaluation of the test run.