hldr-detector: updated HLDR checking algorithms, fixed two bugs introduced before.
Fixed a bug in checkThisViewAgainstOtherHistories() where a wrong window
was acquired (acquired window of the current thread, not the other one).
This bug was accidentally introduced in the previous commit. Fixed a bug
in checkOtherViewsAgainstThisHistory() when two different iterators were
named the same (theoretically it should work fine, but leads to variable
shadowing which is never good).