tools/evaluate: added new option --process-only.
This option forces the evaluation script just to process the results and
do not evaluate them, outputting the processed information for each test
run separately without any summarization at all. This mode is convenient
for arbitrary processing of the results which requires the detailed data
and not only some summaries.