Added the libdie library needed to build the framework as a submodule.
The libdie library was developed mainly to simplify the extraction of
debugging information in the ANaConDA framework. This library may be
used independently, so it is better to keep it in its own repository,
however, it is also convenient to have it available when building the
framework. Adding the library as a submodule let us achieve both.