tools/build: updated to work with the new single-repository layout.
Targets might now be in nested subdirectories. As there is now a clear
hierarchy in the naming of the folders representing the targets (e.g.,
analysers in the analysers folder, libraries in the libraries folder,
etc.), the environment variables' prefixes are now determined
automatically following a set of simple rules.