framework/index: added the INVALID_INDEX special value.
This value is the last value the index might use. If the value is used
when retrieving some data from the index, the program crashes on the
asssertion error. As it is expected that the program never reaches this
value while indexing data, the functions for storing data in the index
currently do not check if the returned value matches the INVALID_INDEX.