tools/cygwin: updated to support execution in shell only (no terminal).
Added the --no-terminal option which forces the script to use bash shell
directly instead of launching it inside the mintty terminal. The command
passing was modified to be done through the new -c option, all arguments
must be enclosed in double quotes or only the first one will be taken by
the -c option.