goodlock: fixed typenames outside of templates, defined throw_exception.
Microsoft Visual C++ compiler does not allow using the typename keyword
outside of templates. Some prior versions of GCC required that in some
cases. As all pintools must be compiled on Windows without exceptions
support, we have to define the throw_exception for Boost, as it does
not define this function without exceptions support.