framework/settings: CODAN cannot analyse some dereferenced iterators.
CODAN cannot infer the directory_entry data type when dereferencing the
directory_iterator and therefore cannot find the methods called on this
type of data thus reporting false alarms. Store the data referenced by
the iterator into a temporary directory_entry& variable and use this
variable to access the data. With this modification, CODAN can analyse
the code and find the methods correctly.