framework/settings: rebinds the analyser to correct ANaConDA framework.
Redirects all calls from the analyser to the ANaConDA framework to the
right (hidden) instance of the ANaConDA framework. This allows us to
use the ANaConDA framework with the latest PIN on Windows, where PIN
is using a custom loader to load the ANaConDA framework DLL and causes
the ANaConDA framework to be invisible to the OS and all DLLs that use
the ANaConDA framework.