tools/run: disable kernel version checks when running kernel 4.0.
The latest version of PIN does not support Linux kernel version 4.0 or
newer. However, there is an undocumented switch which can disable the
kernel version check and thus allows to run PIN on systems with kernel
4.0 and newer. Keep in mind that there is no guarantee that everything
will be working. Still version 4.0 is not that different compared to
the latest version of the 3.0 kernel, so PIN should run just fine.