tools/cygwin: added the --target option for choosing the target arch.
The --target option can be used to configure the cygwin terminal/shell
for building the 32-bit or 64-bit versions of ANaConDA. Each version
requires a different Visual C++ toolchain to be used when building the
framework and its parts and this toolchain needs to be set up before
starting cygwin. By default amd64 toolchain is used on 64-bit systems
and x86 toolchain on 32-bit systems.