tools/build: added the --target-arch option to specify target arch.
By default, target architecture is determined by the operating system,
i.e., if the operating system is 32-bit, a 32-bit version of ANaConDA
is build, if the operating system is 64-bit, a 64-bit version is build.
The --target-arch can be used to override this behaviour and explicitly
specify the target architecture, if it differs from the OS architecture.