tools/cygwin: added support for correcting cygdrive configuration.
The script now detects that the cygdrive simulates Linux permissions
and asks the user if he wants to disable this simulation. When using
the simulation, the framework may not compile successfully. This is
because the folders created by the build script are created under a
Cygwin user account and Cygwin will set the permissions so that this
account may access these folders. However, these folders may also be
accessed by tools from Windows (CMake, Visual Studio tools, etc.)
that access the folders using the Windows user account which may not
have the required permissions to access them. When the simulation is
disabled, folder created from Cygwin inherint the permissions of the
parent folder, i.e., if the user had access to the parent folder, he
will also have access to the created child folder, whether he tries
to access the folder from Cygwin or Windows and that is what we want
to achieve here.