tools/executions: add libdwarf's library directory to LD_LIBRARY_PATH.
The libdwarf library was shipped with PIN in the previous versions so
adding paths to PIN libraries to LD_LIBRARY_PATH also added the path
to the libdwarf library. In the latest versions of PIN, Intel uses a
custom libpindwarf library that consists of parts of the libdwarf and
libelf libraries. This custom library, however, does not contain some
of the functions needed by ANaConDA. That is why we need to use the
standard libdwarf library and thus we need to put the path to it to
the LD_LIBRARY_PATH so it can be found.