tools/build: removed the perform_check function and its external files.
The perform_check function performing various checks using the CMAke was
removed as it is no longer needed by any other function. All the checks
were incorporated into the build script so no external files are needed
anymore. As only the CMake part of the check was in the external files,
the code was split into two places making it hard to manage and update.
Putting all the code into the script makes the management and updating
much easier and it is more clear how is the check performed.