tools/update: added and implemented the '--files tracked' option.
Currently for standard updates only (no --publish). This option updates
all files tracked by the main GIT repository (and its submodules). This
differs from the '--files git' option as it takes the current versions
of these files from the working directory and not the latest versions
from the repository (from the HEAD revision) as does the 'git' option.