tools/executions: split program command into executable and parameters.
When registering a program, one must specify the executable used to run
the program and the parameters passed to this executable separately. It
is easy to construct the full command from these two parts, however, it
is hard to split the full command into the path to the executable and a
list of parameteres because of whitespaces.