tools/build: do not require VS compiler when determining architecture.
If no target architecture is specified, it is derived from the compiler.
However, when preparing the runtime environment, we do not require to
have a Visual Studio installed and thus we may not have the compiler
available to do the check. Added an alternate way to determine the
target architecture (based on the architecture of the OS).