tools/cygwin: added the --quiet option to run with default options.
This option forces the script to automatically take the default choices
and do not prompt the user to make the choice. These choices are mainly
about where to download and install the Cygwin and how to configure it.
In most of the cases, the default options are fine so it is convenient
then to use the --quiet option and take them by default automatically.