framework/thread: fixed function name retrieval in FunctionExited.
Call the OnFunctionExit callbacks before reverting to the information
about the parent function. If we revert the information before calling
the callback functions, we will have wrong information in the callback
functions (information about the parent function and not about the one
we are exiting). Fixed the order of these actions.