framework/thread: threads are now synchronised during thread creation.
When a new thread is created it is properly synchronised with the thread
that created it. This is necessary as both threads have to exchange some
data that has to be prepared before the other thread accesses them. Also
this sychronisation is important for adding support for fork callbacks
where we need to call them when we have both of the threads under our
control (as one may want to access both threads the callback function
and this callback function have to run in one of these threads).