framework/index: changed to implementation to be thread-safe.
This is a temporary solution which fixes a problem that copying strings
is not thread-safe in PIN (because it uses a non-thread-safe version of
the standard library) and compilers use CoW optimisations when copying
strings. These optimisations modify internal pointers present in string
objects and may lead to hidden sharing of parts of the managed string.
This approach is very effective, however, assumes nobody is using the
string concurrently. When other thread is using the same string object,
the behaviour is undefined and may lead to very strange errors (usually
a MmapChecked error in PIN here when we get out-of-bound of the string
when copying it, leading to a situation when we incorrenctly assume the
string is several GiB long and trying to allocate that much memory).