tests: added a new test checking monitoring of locks with backtraces.
In order to provide precise backtraces, the framework needs to monitor
the execution of every function. However, the same thing is needed in
order to detect lock acquisitions and releases. The goal of this test
is to check if monitoring all function calls does not influence the
monitoring of explicitly chosen function calls.