tests: added a new test checking monitoring of functions with longjump.
The goal of the test is to check if the framework reports the correct
number of executions (mainly function exits) of a recursive function
when the function does not return normally but performs a long jump
after a few recursions. The framework should detect this situation
and trigger a correct number of function exit notifications here.