tests: added a new test checking monitoring of the strcmp function.
The goal of the test is to check if the framework reports the correct
number of executions (mainly function enters) of the strcmp function.
This function is often compiled to the assembly code in a way that it
jumps at its beginning after every character comparison. This forces
the framework to trigger the function entered notification more times
that it should be.