framework/index: fixed compilation in release version.
Could not compile the module in release version as some of the variables
were used only in asserts which are ignored in the release version. This
caused the compiler to report an unused variable error which prevented
the module to compile. The variables are now used also in the release
version.