framework/exception: fixed missing function call notification.
When calling an application function, PIN does not actually execute a
CALL instruction which triggers a function call notification. It just
updates the stack to a state in which the stack would be if the CALL
was issued and then resumes the execution. We need to trigger the
notification manually in this case.