tools/build: downloading GCC prerequisites now supports mirror sites.
If the prerequisites cannot be downloaded from the main site, the script
will check all mirror sites and try to retrieve the required files from
there. This may help, e.g., in environments where FTP protocol (used by
the main site) is blocked and thus the files cannot be downloaded from
this location.