tools/update: added a new option --pull for updating the local server.
This new option forces the script to update the local server (itself)
with the files from a remote server (pull mode) instead of updating a
remote server with its local files (push mode). It is designed to be
used with the --publish option which pushes local files onto a remote
server. These files can than be pulled from the remote server into a
different local server to update it.