framework/filter: split tree filter to generic and templated parts.
Split the TreeFilter class into a base class GenericTreeFilter and a
templated class TreeFilter<Data>. The implementation of the tree filter
is not dependant on the custom data available at each node so the base
class uses void* to store this data. To simplify the usage of the base
class, the templated class allows one to specify the type of data that
should be available at each node and handles the convertion between the
void* type and the type specified at a template parameter.