framework/filter: fixed a bug when using found matching path as hint.
When a result of a successfull match is used as a hint, the result
shoult also be a successfull match as we are matching additional
parts of a string sequence with a non-existent nodes which match
any string (they represent a .* regular expression). In this case,
we duplicate the hint and store it as a new result.