framework/settings: arbitrary data can be assigned to hooks, new hook.
Added a new "data function" hook. This hook is not loaded from a hook
file, it is added through an API call instead. The reason is that this
hook extracts an arbitrary data from the parameters of a function. As
the type of the data (and their meaning) is only known to the writer
of the analyser which needs to access them, the callback function must
be tailored to them. So it is natural that such callback function will
be registered in the analyser itself and not through some configuration
file like the other hooks where the data and the semantics of the
functions are known.