Added new analyser, the data validator.
This analyser check the usage of data passed to functions and reports
any attempts to use data that were already disposed of. There exist a
lot of tools for detecting situations where a program tries to use a
NULL pointer, however, these tools usually cannot detect a usage of
data that were disposed of yet still exist in some form. A typical
example is the locks (mutexes) that needs to be initialised before
used and destroyed after they are not needed anymore. However, even
when a lock is destroyed, it often still exist as a local or global
variable so nobody is trying to use invalid (NULL) data. The data are
still valid, however, only as plain data, not as lock representation.
This simple analyser tracks the usage of phtread locks (mutexes) and
reports any attempts to use these locks after they where destroyed.