framework/settings: noise settings in noise points has highest priority.
Load the noise point hooks first as the noise settings defined there has
the highest priority. Because the settings for a specific function can
not be overwritten, this type of hooks has to be loaded first. This is
temporary fix until the hooks can be loaded in a specific order.