framework/thread: report 'call sp != return sp' as warning, not error.
When the SP of some return does not match the SP of the last executed
function, report a warning, ignore the return, and continue. Usually,
every return has a call it belongs to (matches its SP), however, some
low-level libraries contain functions that are never called but still
return after they finish their execution. In most of the cases, the
return is used to move the stack in a desired configuration instead
of modifying the SP directly and ending with a jump to continue the
execution. On 64-bit systems, the direct SP modification combined
with jumps is usually used in these situations, on 32-bit systems,
however, the return is often used instead, causing inconsistency
between the number of calls and returns.