framework/settings: included missing Anaconda API header.
As the settings module contains the implementation of some of the API
functions, it needs to include the anaconda.h header containing the
definition of these API functions. On Linux, these API functions are
exported using an external configuration file, so not including this
header causes no problems. On Windows, however, the API functions are
exported by a directive which is part of the function definition, so
not including the definition causes the function not to be exported.