framework/pe: do not use hints, functions can be sorted in any way.
The call redirection assumed that the functions in import and export
tables were sorted by their names and thus could continue searching
for following functions from the currently found function and not
from the beginning of the list (of functions). Unfortunatelly, this
assumption does not always hold so we need to search the whole list
of functions every time to be sure we redirect all calls correctly.