shared/cmake: use the release version of boost even in debug builds.
The reason is that PIN uses release runtime (CRT library) and the debug
version of boost uses debug runtime. As the release and debug runtimes
cannot be mixed and PIN has no version which uses debug runtime, the
only option here is to use a release version of boost in all builds.