framework/pe: included additional header to solve some indexer errors.
The CODAN indexer (from Eclipse CDT) has problems indexing some defines
in the winnt.h header, which is very large. This header is included by
a long chain of header files and it seems the indexer ignores some of
the deeply nested header files. By including this header file in the
source file itself, the indexer resolves all symbols correctly.