framework/thread: extended the API functions for monitoring functions.
Added a new variant of the THREAD_FunctionExecuted API function which
takes two callback functions instead of one. While the first function
is called when a specific function starts its execution, the second
one is called when the function finishes its execution. The first
function can access the data of one of the arguments of the executed
function. The second function can access the data returned by the
executed function.