tools/build: do not require libdwarf on mac for now.
We cannot compile libdwarf without libelf, and libelf is not used on Mac
OS X, so do not compile the library for now. It will be needed later on,
however we will add it after it is known which parts are required on mac
and which parts can be safely ignored.