tools/build: build GCC 4.9.4 when no usable (compatible) GCC is found.
ANaConDA cannot be compiled with newer GCC compilers that support the
Dual ABI. Because of PIN, which is compiled with the old ABI, we can
only compile ANaConDA with the same (old) ABI. However, we cannot use
the libstdc++ library even if it supports both old and new ABI (Dual
ABI) as programs using the old ABI crash randomly (mainly std::string
operations cause the crashes in the form of seqfaults). The only way
to compile the ANaConDA framework is to use older compiler which does
not support the newer ABI at all. As GCC 4.9.3 cannot be compiled by
the newer GCC compilers anymore, we have to use GCC 4.9.4 now, which
contains patches that allows it to be compiled by newer GCC compilers
(5.x, 6.x, ...) and supports the old ABI only.