anaconda: fixed function name retrieval in FunctionExecuted callbacks.
The GetCurrentFunction() helper function was returning the name of the
parent function when used in the function executed callback function.
This was because this callback function was called before the function
entered callback function, which updates the name of the currently
executed function. Fixed by forcing PIN to call the function entered
callback function first.