tools/build: check exit status of CMake when localizing libdwarf/libelf.
Do not determine if libdwarf or libelf was found just by parsing the
output of CMake. Check the exit status of CMake to determine if the
search was successful or not. Only when the search was successful,
extract the information needed from the output.

If the script is unable to find any of the libraries, one can use the
--verbose option get the output of CMake and see why it failed.