framework/cbstack: added detection of previously missed return(s).
When a return is encountered, check if the call stack does not contain
registered callback functions for stack frames already destroyed. This
situation may happen if a return was missed during the monitoring.

Ideally, such situation never happens, however, if a return is missed,
the registered callback function is at least triggered when the first
return after the missed one is encountered.