framework/anaconda: expanded memory access instrumentation macros.
Most of these macros are used only once, so there is no point of having
them and they obscure now the memory access instrumentation is done. In
order to redesign the memory access instrumentation, they were expanded
to show more clearly how the instrumentation is done and how it can be
redesigned to make it better.