tools/run: use functions to run constructed commands instead of eval.
The main goal of the run script is to construct a command to perform an
analysis or to execute a program directly and then run the command. The
script was using `eval` to run the command. While using `eval` for this
task is convenient in many ways, it may also be dangerous and cause all
sorts of issues with handling parameters etc. Using functions to create
(define) the commands to run is much more cleaner and safer. As part of
the changes, various arguments are now also handled as lists.