tools/run: updated debug support to work with new command run system.
Moved parts of the script configuring the analysis or program to run in
debug mode to a separate function. Updated the function to be compatible
with the new way of running the constructed commands using functions.

Changed the format of ANACONDA_FLAGS to a list like PIN_FLAGS. Updated
the functions using these variables to handle them as lists.