goodlock/cycles: modified cycle enumeration to use handlers for cycles.
Using handlers for handling cycles when they are found is more generic
than restricting the implementation to collect all cycles to a list.
Added a handler that collects all cycles to a list to provide the old
behavior while allowing to override this behavior and define your own
way of handling cycles when they are found by the cycle enumeration.