goodlock: added support for printing potential deadlocks immediately.
Updated the configuration entry 'show.deadlocks' to take one of the
following values: never, immediately, and finally. The immediately
option can be used to print potential deadlocks immediately during
the enumeration of cycles in the lock graph. The finally option is
the old behavior where all cycles are enumerated first and printed
after it. This may force the analyser to store a lot of cycles that
are not valid and cannot cause a deadlock. The new behavior is the
default now.