tools/executions: prefer the libelf library used to compile ANaConDA.
ANaConDA requires several internal header files from the libelf library
to compile. These files are not included in most of the libelf packages
present in various linux distributions, therefore we are usually forced
to build the library ourselves and use it to compile ANaConDA. Similar
to the libdwarf library, we should use the same same libelf library we
used to compile ANaConDA for its runtime.